@@ -1,54 +1,64 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
-import { LandingPageComponent } from './views/landing-page/landing-page.component';
+import { LandingPageComponent } from './modules/landing-page/view/landing-page.component';
+import { CheckToken } from './shared/guards/checkToken.guard';
 
 const routes: Routes = [
   {
     path: '',
-    component: LandingPageComponent
+    component: LandingPageComponent,
+    canActivate: [CheckToken],
   },
   {
     path: 'auth',
-    loadChildren: () => import('./views/auth/auth-routing.module').then(m => m.AuthRoutingModule)
+    loadChildren: () =>
+      import('./modules/auth/auth-routing.module').then(
+        (m) => m.AuthRoutingModule
+      ),
   },
-  { 
-    path: 'info',
-    loadChildren: () => import('./views/info/info-routing.module').then(m => m.InfoRoutingModule)
-  },
-  { 
+  {
     path: 'basic-cars',
-    loadChildren: () => import('./views/basic-cars-page/basic-cars-routing.module').then(m => m.BasicCarsRoutingModule)
+    loadChildren: () =>
+      import('./modules/basic-cars/basic-cars-routing.module').then(
+        (m) => m.BasicCarsRoutingModule
+      ),
   },
-  { 
+  {
     path: 'custom-cars',
-    loadChildren: () => import('./views/custom-cars/custom-cars-routing.module').then(m => m.CustomCarsRoutingModule)
+    loadChildren: () =>
+      import('./modules/custom-cars/custom-cars-routing.module').then(
+        (m) => m.CustomCarsRoutingModule
+      ),
   },
-  { 
-    path: 'detailed-car/:type/:id',
-    loadChildren: () => import('./views/detailed-car-page/detailed-car-routing.module').then(m => m.DetailedCarRoutingModule)
-  },
-  { 
+  {
     path: 'premium-cars',
-    loadChildren: () => import('./views/premium-cars-page/premium-cars-routing.module').then(m => m.PremiumCarsRoutingModule)
+    loadChildren: () =>
+      import('./modules/premium-cars/premium-cars-routing.module').then(
+        (m) => m.PremiumCarsRoutingModule
+      ),
   },
-  { 
+  {
     path: 'search/:query',
-    loadChildren: () => import('./views/search-results/search-results-routing.module').then(m => m.SearchResultsRoutingModule)
+    loadChildren: () =>
+      import('./modules/search-results/search-results-routing.module').then(
+        (m) => m.SearchResultsRoutingModule
+      ),
   },
-  { 
+  {
     path: 'user',
-    loadChildren: () => import('./views/user/user-routing.module').then(m => m.UserRoutingModule)
+    loadChildren: () =>
+      import('./modules/user/user-routing.module').then(
+        (m) => m.UserRoutingModule
+      ),
   },
   {
     path: '**',
-    redirectTo: ''
-  }
-  
+    redirectTo: '',
+  },
 ];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class AppRoutingModule { }
+export class AppRoutingModule {}
