@@ -1,12 +1,10 @@
-import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Observable, from } from 'rxjs';
+import { Injectable } from '@angular/core';
+import { Observable, from, of } from 'rxjs';
 import { mergeMap } from 'rxjs/operators';
 import { userInterfaceApi, userUpdateRequest } from '../models/user.interface';
 
 import { environment } from 'src/environments/environment';
-import { getTokenFromIndexedDB } from '../helpers/indexedDB';
-
 
 @Injectable({
   providedIn: 'root',
@@ -15,89 +13,107 @@ export class UserService {
   constructor(private http: HttpClient) {}
 
   async getUserData(id: number): Promise<userInterfaceApi | any> {
-    const response = await this.http.get(`${environment.apiBaseUrl}/api/users/${id}`).toPromise();
+    const response = await this.http
+      .get(`${environment.apiBaseUrl}/api/users/${id}`)
+      .toPromise();
     return response as userInterfaceApi;
   }
 
   async getUserByUsername(username: string): Promise<userInterfaceApi | any> {
-    const response = await this.http.get(`${environment.apiBaseUrl}/api/users/username/${username}`).toPromise();
+    const response = await this.http
+      .get(`${environment.apiBaseUrl}/api/users/username/${username}`)
+      .toPromise();
     return response as userInterfaceApi;
-  }  
+  }
 
   updateUser(id_user: number, userParams: userUpdateRequest): Observable<any> {
-    return from(getTokenFromIndexedDB()).pipe(
-      mergeMap(token => {
-        const headers = new HttpHeaders({
-          'Content-Type': 'application/json',
-          'r-token': token
-        });
-  
-        const requestBody: userUpdateRequest = {}
-  
-        if (userParams.username) {
-            requestBody.username = userParams.username;
-        }
-        if (userParams.email) {
-            requestBody.email = userParams.email;
-        }
-        if (userParams.password) {
-            requestBody.password = userParams.password;
-        }
-  
-        return this.http.put(`${environment.apiBaseUrl}/api/users/${id_user}`, requestBody, { headers });
-      })
-    );
+    const token = localStorage.getItem('dt-token');
+    if (token) {
+      const headers = new HttpHeaders({
+        'Content-Type': 'application/json',
+        'r-token': token,
+      });
+
+      const requestBody: userUpdateRequest = {};
+
+      if (userParams.username) {
+        requestBody.username = userParams.username;
+      }
+      if (userParams.email) {
+        requestBody.email = userParams.email;
+      }
+      if (userParams.password) {
+        requestBody.password = userParams.password;
+      }
+
+      return this.http.put(
+        `${environment.apiBaseUrl}/api/users/${id_user}`,
+        requestBody,
+        { headers }
+      );
+    } else return of(false);
   }
 
   updateImg(id_user: number, file: File, bg: boolean = false): Observable<any> {
-    return from(getTokenFromIndexedDB()).pipe(
-      mergeMap(token => {
-        const headers = new HttpHeaders({ 'r-token': token! });
-        const formData = new FormData();
-        
-        formData.append('file', file);
-      
-        if (bg) {
-          return this.http.post(`${environment.apiBaseUrl}/api/users/bg-img/${id_user}`, formData, { headers });
-        }
-        
-        return this.http.post(`${environment.apiBaseUrl}/api/users/img/${id_user}`, formData, { headers });
-      })
+    const token = localStorage.getItem('dt-token');
+    const headers = new HttpHeaders({ 'r-token': token! });
+    const formData = new FormData();
+
+    formData.append('file', file);
+
+    if (bg) {
+      return this.http.post(
+        `${environment.apiBaseUrl}/api/users/bg-img/${id_user}`,
+        formData,
+        { headers }
+      );
+    }
+
+    return this.http.post(
+      `${environment.apiBaseUrl}/api/users/img/${id_user}`,
+      formData,
+      { headers }
     );
   }
 
   deleteImg(id_user: number, bg: boolean = false): Observable<any> {
-    return from(getTokenFromIndexedDB()).pipe(
-      mergeMap(token => {
-        const headers = new HttpHeaders({ 'r-token': token! });    
-      
-        if (bg) {
-          return this.http.delete(`${environment.apiBaseUrl}/api/users/bg-img/${id_user}`, { headers });
-        }
-        
-        return this.http.delete(`${environment.apiBaseUrl}/api/users/img/${id_user}`, { headers });
-      }));
+    const token = localStorage.getItem('dt-token');
+    const headers = new HttpHeaders({ 'r-token': token! });
+
+    if (bg) {
+      return this.http.delete(
+        `${environment.apiBaseUrl}/api/users/bg-img/${id_user}`,
+        { headers }
+      );
+    }
+
+    return this.http.delete(
+      `${environment.apiBaseUrl}/api/users/img/${id_user}`,
+      { headers }
+    );
   }
 
   async downloadUserCollection(id: number): Promise<void> {
     try {
-      const token = await getTokenFromIndexedDB();
+      const token = localStorage.getItem('dt-token');
+      if (token) {
+        const headers = new HttpHeaders({ 'r-token': token });
 
-      const headers = new HttpHeaders({ 'r-token': token });
+        const response = await this.http
+          .get(`${environment.apiBaseUrl}/api/users/get-collection/${id}`, {
+            responseType: 'blob',
+            headers: headers,
+          })
+          .toPromise();
 
-      const response = await this.http.get(`${environment.apiBaseUrl}/api/users/get-collection/${id}`, { 
-          responseType: 'blob',
-          headers: headers 
-      }).toPromise();
-
-      if (response) {
+        if (response) {
           this.triggerDownload(response, `user_${id}_collection.zip`);
+        }
       }
     } catch (error: any) {
-        throw new Error('Error downloading the user collection');
+      throw new Error('Error downloading the user collection');
     }
   }
-  
 
   private triggerDownload(data: Blob, filename: string): void {
     const blob = new Blob([data], { type: 'text/csv' });
@@ -109,5 +125,4 @@ export class UserService {
     link.click();
     link.remove();
   }
-  
 }
