@@ -1,139 +1,127 @@
-import { Component, OnInit, EventEmitter, Output } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
-import { BasicCarsService } from '../../basic-cars-page/basic-cars.service';
-import { PremiumCarsService } from '../../premium-cars-page/premium-cars.service';
-import { UserService } from '../../../services/user.service';
-import { userInterface } from 'src/app/models/user.interface';
-import { basicCarInterface, basicCarsGrouped, basicGlobalGroup, customCarInterface, premiumCarInterface, premiumCarsGrouped, premiumGlobalGroup } from 'src/app/models/cardTypes.interface';
-import { decodeToken } from 'src/app/helpers/generics';
-import { CustomCarsService } from '../../custom-cars/custom-cars.service';
-import { LoaderService } from 'src/app/services/loader.service';
-import { createSpecialGroup, filterMainSerie, filterSeries, filterYear } from 'src/app/helpers/filter-series';
+import { Component } from '@angular/core';
+import { UserService } from 'src/app/services/user.service';
 
 @Component({
   selector: 'app-user-profile',
   templateUrl: './user-profile.component.html',
-  styleUrls: ['./user-profile.component.css', '../../../styles/cars-views.css']
+  styleUrls: ['./user-profile.component.css'],
 })
-export class UserProfileComponent implements OnInit {
-
+export class UserProfileComponent {
   @Output() errorEvent = new EventEmitter<string>();
-
   user?: userInterface;
   userVisitor = false;
-
   // Tipo de coche seleccionado
   carsTypeSelected = 'basic';
   collectedTypeSelected = 'owned';
-
   basicCarsOwned: basicGlobalGroup[] = []; // Básicos en propiedad
   basicCarsWished: basicGlobalGroup[] = []; // Básicos deseados
   basicCarsOwnedNumber: number = 0;
   carsGroupedSeries: number = 0; // Número de grupos básicos (Sirve para el grupo especial en la última posicion + 1)
-
   premiumCarsOwned: premiumGlobalGroup[] = []; // Premium en propiedad
   premiumCarsWished: premiumGlobalGroup[] = []; // Premium deseados
   premiumCarsOwnedNumber: number = 0;
-
   customCarsOwned = []; // Customs en propiedad
-
   // Filtro seleccionado y disponibles (Basic cars)
   selectedYear = 'ALL';
   selectedSerie = '';
-  availableYears = ['2024', '2023', '2022', '2021', '2020', '2019', '2018', '2017', '2016'];
+  availableYears = [
+    '2024',
+    '2023',
+    '2022',
+    '2021',
+    '2020',
+    '2019',
+    '2018',
+    '2017',
+    '2016',
+  ];
   availableSeries = [];
-
   // Filtro seleccionado y disponibles (Premium cars)
   selectedMainSerie = 'ALL';
   selectedSecondarySerie = '';
-  availableMainSeries = ['Boulevard (original)', 'Boulevard (reboot)', 'Car Culture', 'Fast & Furious (Premium)', 'Fast & Furious', 'Pop Culture'];
+  availableMainSeries = [
+    'Boulevard (original)',
+    'Boulevard (reboot)',
+    'Car Culture',
+    'Fast & Furious (Premium)',
+    'Fast & Furious',
+    'Pop Culture',
+  ];
   availableSecondarySeries = [];
-
   // Series especiales (Mover a BE)
-  specialSeries = ['Treasure Hunt', 'Super Treasure Hunt', 'Walmart Exclusive', 'Kroger Exclusive'];
-
+  specialSeries = [
+    'Treasure Hunt',
+    'Super Treasure Hunt',
+    'Walmart Exclusive',
+    'Kroger Exclusive',
+  ];
   // Error handlers
   error = false;
   errorMsg = '';
   exportToCsvError = false;
-
   constructor(
     private basicCarsService: BasicCarsService,
     private customCarsService: CustomCarsService,
-    private loaderService: LoaderService,
     private premiumCarsService: PremiumCarsService,
     private route: ActivatedRoute,
     private router: Router,
-    private userService: UserService,
-  ) { }
-
- ngOnInit() {
-  this.loaderService.startLoading();
-  
-  this.route.paramMap.subscribe(async params => {
-    const username = params.get('username');
-
-    if (username) {
-      try {
-        const userResponse = await this.userService.getUserByUsername(username);
-        this.user = userResponse.user;
-  
-        const tokenDecoded = await decodeToken();
-        const isUserOwner = tokenDecoded.userId === this.user?.id;
-        this.userVisitor = isUserOwner;
-  
-        if (this.user) {
-          this.getCustomCars(this.user.id);
-  
-          try {
-            const [basicCarsResponse, premiumCarsResponse] = await Promise.all([
-              this.basicCarsObservable().toPromise(),
-              this.premiumCarsObservable().toPromise()
-            ]);
-
-            this.processBasicCars(basicCarsResponse, isUserOwner);
-            this.processPremiumCars(premiumCarsResponse, isUserOwner);
-          } catch (err) {
-            console.error(err);
+    private userService: UserService
+  ) {}
+  ngOnInit() {
+    this.route.paramMap.subscribe(async (params) => {
+      const username = params.get('username');
+      if (username) {
+        try {
+          const userResponse = await this.userService.getUserByUsername(
+            username
+          );
+          this.user = userResponse.user;
+          const tokenDecoded = await decodeToken();
+          const isUserOwner = tokenDecoded.userId === this.user?.id;
+          this.userVisitor = isUserOwner;
+          if (this.user) {
+            this.getCustomCars(this.user.id);
+            try {
+              const [basicCarsResponse, premiumCarsResponse] =
+                await Promise.all([
+                  // this.basicCarsObservable().toPromise(),
+                  this.premiumCarsObservable().toPromise(),
+                ]);
+              this.processBasicCars(basicCarsResponse, isUserOwner);
+              this.processPremiumCars(premiumCarsResponse, isUserOwner);
+            } catch (err) {
+              console.error(err);
+            }
           }
+        } catch (error) {
+          console.error('User not found:', error);
+          this.error = true;
+          this.router.navigate(['/']);
+        } finally {
         }
-      } catch (error) {
-        console.error('User not found:', error);
+      } else {
         this.error = true;
         this.router.navigate(['/']);
-      } finally {
-        this.loaderService.stopLoading();
       }
-    } else {
-      this.error = true;
-      this.router.navigate(['/']);
-      this.loaderService.stopLoading();
-    }
-  });
-  
- }
-
+    });
+  }
   processBasicCars(res: any, isUserOwner: boolean) {
     const transformGroup = (group: basicGlobalGroup) => {
       group.series = group.series.map((serie: any) => {
         serie.cars = serie.cars.map(transformCar);
-
         return {
           ...serie,
-          visible: true
-        }
+          visible: true,
+        };
       });
-      
       return {
         ...group,
-        visible: true
-      }    
-    }
-  
+        visible: true,
+      };
+    };
     const transformCar = (car: basicCarInterface) => {
       const series = car.series.split(',');
       const serie_class = series[0].replace(/ /g, '-').toLowerCase();
-  
       return {
         ...car,
         series,
@@ -143,38 +131,32 @@ export class UserProfileComponent implements OnInit {
         profile_view: true,
         has_car: isUserOwner,
         token: isUserOwner ? this.user!.id : undefined,
-        visible: true
+        visible: true,
       };
     };
-  
     this.basicCarsOwned = res.groupedOwnedCars.map(transformGroup);
     this.basicCarsWished = res.groupedWishedCars.map(transformGroup);
-
     // Contar el número de coches básicos que tiene el usuario
     this.basicCarsOwned.forEach((group: basicGlobalGroup) => {
       group.series.forEach((serie: basicCarsGrouped) => {
         serie.cars.forEach(() => this.basicCarsOwnedNumber++);
-      })
+      });
     });
   }
-  
   processPremiumCars(res: any, isUserOwner: boolean) {
     const transformGroup = (group: premiumGlobalGroup) => {
       group.secondarySeries = group.secondarySeries.map((serie: any) => {
         serie.cars = serie.cars.map(transformCar);
-
         return {
           ...serie,
-          visible: true
-        }
+          visible: true,
+        };
       });
-      
       return {
         ...group,
-        visible: true
-      }    
-    }
-  
+        visible: true,
+      };
+    };
     const transformCar = (car: premiumCarInterface) => {
       return {
         ...car,
@@ -182,13 +164,11 @@ export class UserProfileComponent implements OnInit {
         profile_view: true,
         has_car: isUserOwner,
         token: isUserOwner ? this.user!.id : undefined,
-        visible: true
+        visible: true,
       };
     };
-  
     this.premiumCarsOwned = res.groupedOwnedCars.map(transformGroup);
     this.premiumCarsWished = res.groupedWishedCars.map(transformGroup);
-
     // Contar el número de coches premium que tiene el usuario
     this.premiumCarsOwned.forEach((group: premiumGlobalGroup) => {
       group.secondarySeries.forEach((serie: premiumCarsGrouped) => {
@@ -196,15 +176,12 @@ export class UserProfileComponent implements OnInit {
       });
     });
   }
-
   basicCarsObservable() {
-    return this.basicCarsService.getUserCars(this.user!.id);
+    // return this.basicCarsService.getUserCars(this.user!.id);
   }
-  
   premiumCarsObservable() {
     return this.premiumCarsService.getUserCars(this.user!.id);
   }
-
   enableErrorMsg(msg: string | any) {
     this.error = true;
     if (typeof msg === 'string') {
@@ -213,26 +190,19 @@ export class UserProfileComponent implements OnInit {
       this.errorMsg = 'GN_UNEXPECTED_ERROR';
     }
   }
-
   changeCarType(type: string) {
     this.carsTypeSelected = type;
-
-    if(type === 'custom') {
-      this.collectedTypeSelected = 'owned'
+    if (type === 'custom') {
+      this.collectedTypeSelected = 'owned';
     }
   }
-
   changeCarCollected(type: string) {
     this.collectedTypeSelected = type;
   }
-
   filterBasicYear(year: string) {
     this.selectedYear = year;
     this.selectedSerie = 'ALL';
-
     year !== 'ALL' ? this.getAvailableSeries(year) : null;
-    
-
     if (year === 'ALL') {
       this.basicCarsOwned = filterYear(this.basicCarsOwned);
       this.basicCarsWished = filterYear(this.basicCarsWished);
@@ -241,43 +211,42 @@ export class UserProfileComponent implements OnInit {
       this.basicCarsWished = filterYear(this.basicCarsWished, year);
     }
   }
-
   filterBasicSeries(serie: string) {
     this.selectedSerie = serie;
-
     if (this.specialSeries.includes(serie)) {
-      this.basicCarsOwned.forEach(group => {
-        group.series = createSpecialGroup(group.series, serie, this.carsGroupedSeries);
+      this.basicCarsOwned.forEach((group) => {
+        group.series = createSpecialGroup(
+          group.series,
+          serie,
+          this.carsGroupedSeries
+        );
       });
-
-      this.basicCarsWished.forEach(group => {
-        group.series = createSpecialGroup(group.series, serie, this.carsGroupedSeries);
+      this.basicCarsWished.forEach((group) => {
+        group.series = createSpecialGroup(
+          group.series,
+          serie,
+          this.carsGroupedSeries
+        );
       });
-
     } else {
       this.basicCarsOwned.forEach((group: basicGlobalGroup) => {
         group.series = filterSeries(group.series, serie);
       });
-
       this.basicCarsWished.forEach((group: basicGlobalGroup) => {
         group.series = filterSeries(group.series, serie);
       });
     }
   }
-
   getAvailableSeries(year: string) {
-    this.basicCarsService.getAvailableSeries(year).subscribe(res => {
-      const series = res.series.split(',');
-      this.availableSeries = series.sort();
-    })
+    this.basicCarsService.getAvailableSeries(year).subscribe((res) => {
+      // const series = res.series.split(',');
+      // this.availableSeries = series.sort();
+    });
   }
-
   filterPremiumSeries(serie: string) {
     this.selectedMainSerie = serie;
     this.selectedSecondarySerie = 'ALL';
-
     serie !== 'ALL' ? this.getAvailablePremiumSeries(serie) : null;
-
     if (serie === 'ALL') {
       this.premiumCarsOwned = filterMainSerie(this.premiumCarsOwned);
       this.premiumCarsWished = filterMainSerie(this.premiumCarsWished);
@@ -286,25 +255,21 @@ export class UserProfileComponent implements OnInit {
       this.premiumCarsWished = filterMainSerie(this.premiumCarsWished, serie);
     }
   }
-
   filterPremiumSecundarySeries(serie: string) {
     this.premiumCarsOwned.forEach((group: premiumGlobalGroup) => {
       group.secondarySeries = filterSeries(group.secondarySeries, serie);
     });
-
     this.premiumCarsWished.forEach((group: premiumGlobalGroup) => {
       group.secondarySeries = filterSeries(group.secondarySeries, serie);
     });
-    console.log(this.premiumCarsOwned)
+    console.log(this.premiumCarsOwned);
   }
-
   getAvailablePremiumSeries(main_serie: string) {
-    this.premiumCarsService.getAvailableSeries(main_serie).subscribe(res => {
+    this.premiumCarsService.getAvailableSeries(main_serie).subscribe((res) => {
       const series = res.series.split(',');
       this.availableSecondarySeries = series.sort();
-    })
+    });
   }
-
   getCustomCars(userCreator: number) {
     this.customCarsService.getCustomCars(userCreator).subscribe(
       (res) => {
@@ -314,7 +279,6 @@ export class UserProfileComponent implements OnInit {
             imgs: car.imgs.split(','),
           };
         });
-
         this.customCarsOwned = customCars;
       },
       (err: string) => {
@@ -322,20 +286,15 @@ export class UserProfileComponent implements OnInit {
       }
     );
   }
-
   goToConfig() {
-    this.loaderService.startLoading();
     this.router.navigate(['/user/config']);
   }
-
   exportToCsv() {
     if (this.user) {
-      this.userService.downloadUserCollection(this.user.id)
-        .catch(err => {
-          console.error(err);
-          this.exportToCsvError = true;
-        });
+      this.userService.downloadUserCollection(this.user.id).catch((err) => {
+        console.error(err);
+        this.exportToCsvError = true;
+      });
     }
   }
-
 }
